Remove run-dataset command, integrate eval support into run-prompt
